Make Adapt and InverseFunctions hard dependencies again
As package extensions they can deadlock parallel precompilation on
Julia 1.10: environments that load ValueShapes together with Adapt,
InverseFunctions and further packages hang reproducibly, with the
precompile workers idle and the coordinator holding the extension
pidfiles (Julia 1.11 and later are not affected). Both are tiny
interface packages, so making them regular dependencies again (as
before v0.11.5) costs practically nothing.

Created by generative AI.
