+++ Project.toml
@@ -1,14 +1,16 @@
 name = "ValueShapes"
 uuid = "136a8f8c-c49b-4edb-8b98-f3d64d48be8f"
-version = "0.11.5"
+version = "0.11.6"
 
 [deps]
+Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 ArgCheck = "dce04be8-c92d-5529-be00-80e4d2c0e197"
 ArraysOfArrays = "65a8f2f4-9b39-5baf-92e2-a9cc46fdf018"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ElasticArrays = "fdbdab4c-e67f-52f5-8c3f-e7b388dad3d4"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
+InverseFunctions = "3587e190-3f89-42d0-90ee-14403ec27112"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
@@ -16,18 +18,14 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 TypedTables = "9d95f2ec-7b3d-5a63-8d20-e2491e220bb9"
 
 [weakdeps]
-Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 ChangesOfVariables = "9e997f8a-9a97-42d5-a9f1-ce6bfc15e2c0"
-InverseFunctions = "3587e190-3f89-42d0-90ee-14403ec27112"
 Mooncake = "da2b9cff-9c12-43a0-ae48-6db2b0edb7d6"
 ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
 
 [extensions]
-ValueShapesAdaptExt = "Adapt"
 ValueShapesChainRulesCoreExt = "ChainRulesCore"
 ValueShapesChangesOfVariablesExt = "ChangesOfVariables"
-ValueShapesInverseFunctionsExt = "InverseFunctions"
 ValueShapesMooncakeExt = "Mooncake"
 ValueShapesZygoteRulesExt = "ZygoteRules"
